Farmbot Cloud/Classificatie-container/requirements.txt
flask==2.0.2
gunicorn==20.1.0
google-cloud-storage==1.44.0